M13: split Assistant::InputBuilder into per-concern submodules
Replaces the temporary Metrics/ModuleLength: 150 override (added in M7)
with a structural split along the responsibilities already implicit in
the file:

  lib/assistant/input_builder/
    registry.rb           # Registry        (input_definitions, register_…)
    default_option.rb     # DefaultOption   (M1 trio)
    optional_option.rb    # OptionalOption  (M7 trio)
    accessors.rb          # Accessors       (getter + checker; refinement-scoped here)
    require_validator.rb  # RequireValidator
    type_validator.rb     # TypeValidator   (M3)
    dsl.rb                # Dsl             (input, inputs)

The umbrella lib/assistant/input_builder.rb requires + includes each
submodule. Service still 'extend Assistant::InputBuilder'; no public
API change.

The using Assistant::Refinements::StringBlankness directive narrows
from module-wide to just accessors.rb (lexically scoped where it is
actually consumed).

Tests mirror the lib layout under test/assistant/input_builder/.
The umbrella test/assistant/input_builder_test.rb keeps only
structural smoke tests: every submodule is included, every helper
method is reachable from a fresh extend, and an end-to-end DSL
sanity check. The capture_io_warn helper relocates to
TestHelpers::IoCapture in test/test_helper.rb.

Removes Metrics/ModuleLength block from .rubocop.yml; the umbrella
is well under the default 100-line ceiling.

Adds M13 to docs/v1/02-features.md (Must list, lands before M12) and
a cross-reference under RuboCop in docs/v1/05-quality-and-tooling.md.

Suite: 112 runs / 318 assertions / 0 failures (was 86/249).
RuboCop: clean against default ceilings across 30 files.
